requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.9.0
anthropic>=0.40.0
aiofiles>=24.0
python-multipart>=0.0.9
httpx>=0.27.0
openai>=1.50.0
